Add files via upload
Are you a passionate web developer with a flair for design? Do you have expertise in HTML, CSS, and JavaScript? If so, we have an amazing opportunity for you!

We are looking for a skilled developer to join our team and help us create a cutting-edge hotel management website. As part of our team, you will have the chance to showcase your creativity and technical skills while contributing to a project that will revolutionize the hospitality industry.

Key Responsibilities:
🔹 Designing and developing a user-friendly hotel management website using HTML, CSS, and JavaScript.